feat: Add YouTube video field to ideas
Introduces a new `youtube_video_url` field to the `Idea` type. This allows users to associate a standard YouTube video with their ideas, separate from showroom videos. The `IdeaCard` and `IdeaDetailModal` components have been updated to display a YouTube icon on cards with videos and to embed videos in the detail view.